add order customer func index 08/12/2018

+++ src/main/java/edu/poly/controller/ManagementController.java
@@ -98,12 +98,11 @@ public class ManagementController {
         ModelAndView mav = new ModelAndView();
 
         if (!CheckSession.isLogin(session)) {
-            mav.setViewName(INDEX_SCREEN);
+            mav.setViewName("redirect:/" + Constants.Characters.BLANK);
             return mav;
         }
 
         Users users = (Users) session.getAttribute(Constants.SessionKey.USER);
-
         mav.addObject("userInfo", users);
         mav.setViewName(INFOMATION_SCREEN);
         return mav;
@@ -113,7 +112,7 @@ public class ManagementController {
     public ModelAndView managementTour(HttpSession session) {
         ModelAndView mav = new ModelAndView();
 
-        if (!CheckSession.partner(session)) {
+        if (!CheckSession.isLogin(session)) {
             mav.setViewName("redirect:/" + Constants.Characters.BLANK);
             return mav;
         }
@@ -130,7 +129,7 @@ public class ManagementController {
     @GetMapping(Constants.Url.MANAGEMENT_DELETE_ORDER_URL)
     public ModelAndView managementDeleteOrder(HttpSession session) {
         ModelAndView mav = new ModelAndView();
-        if (!CheckSession.partner(session)) {
+        if (!CheckSession.isLogin(session)) {
             mav.setViewName("redirect:/" + Constants.Characters.BLANK);
             return mav;
         }
@@ -151,7 +150,7 @@ public class ManagementController {
     @GetMapping(Constants.Url.MANAGEMENT_PURCHASE_ORDER_URL)
     public ModelAndView managementPurchaseOrder(HttpSession session) {
         ModelAndView mav = new ModelAndView();
-        if (!CheckSession.partner(session)) {
+        if (!CheckSession.isLogin(session)) {
             mav.setViewName("redirect:/" + Constants.Characters.BLANK);
             return mav;
         }
@@ -172,7 +171,7 @@ public class ManagementController {
     @GetMapping(Constants.Url.MANAGEMENT_ORDER_URL)
     public ModelAndView managementOrder(HttpSession session) {
         ModelAndView mav = new ModelAndView();
-        if (!CheckSession.partner(session)) {
+        if (!CheckSession.isLogin(session)) {
             mav.setViewName("redirect:/" + Constants.Characters.BLANK);
             return mav;
         }
@@ -193,10 +192,10 @@ public class ManagementController {
     @GetMapping(Constants.Url.MANAGEMENT_PARTNER_ORDER_URL)
     public ModelAndView managementPartnerOrder(HttpSession session) {
         ModelAndView mav = new ModelAndView();
-//        if (!CheckSession.partner(session)) {
-//            mav.setViewName("redirect:/" + Constants.Characters.BLANK);
-//            return mav;
-//        }
+        if (!CheckSession.partner(session)) {
+            mav.setViewName("redirect:/" + Constants.Characters.BLANK);
+            return mav;
+        }
         try {
             Users users = (Users) session.getAttribute(Constants.SessionKey.USER);
             orderDAO.deleteOrderAdmin();
@@ -214,10 +213,10 @@ public class ManagementController {
     @GetMapping(Constants.Url.MANAGEMENT_PARTNER_DELETE_ORDER_URL)
     public ModelAndView managementPartnerDeleteOrder(HttpSession session) {
         ModelAndView mav = new ModelAndView();
-//        if (!CheckSession.partner(session)) {
-//            mav.setViewName("redirect:/" + Constants.Characters.BLANK);
-//            return mav;
-//        }
+        if (!CheckSession.partner(session)) {
+            mav.setViewName("redirect:/" + Constants.Characters.BLANK);
+            return mav;
+        }
         try {
             Users users = (Users) session.getAttribute(Constants.SessionKey.USER);
             orderDAO.deleteOrderAdmin();
@@ -235,10 +234,10 @@ public class ManagementController {
     @GetMapping(Constants.Url.MANAGEMENT_PARTNER_PURCHASE_ORDER_URL)
     public ModelAndView managementPartnerPurchaseOrder(HttpSession session) {
         ModelAndView mav = new ModelAndView();
-//        if (!CheckSession.partner(session)) {
-//            mav.setViewName("redirect:/" + Constants.Characters.BLANK);
-//            return mav;
-//        }
+        if (!CheckSession.partner(session)) {
+            mav.setViewName("redirect:/" + Constants.Characters.BLANK);
+            return mav;
+        }
         try {
             Users users = (Users) session.getAttribute(Constants.SessionKey.USER);
             orderDAO.deleteOrderAdmin();
@@ -257,13 +256,18 @@ public class ManagementController {
     @GetMapping(Constants.Url.PURCHASE_PARTNER_ORDER)
     public ModelAndView purchaseOrder(HttpSession session, @PathVariable("id") Integer id) {
         ModelAndView mav = new ModelAndView();
-//        if (!CheckSession.partner(session)) {
-//            mav.setViewName("redirect:/" + Constants.Characters.BLANK);
-//            return mav;
-//        }
+        if (!CheckSession.partner(session)) {
+            mav.setViewName("redirect:/" + Constants.Characters.BLANK);
+            return mav;
+        }
         try {
-            orderDAO.updatePurchaseOrder(true, id);
             OrderDTO orderDTO = orderDAO.getOrderDTO(id);
+            Users users = (Users) session.getAttribute(Constants.SessionKey.USER);
+            if(orderDTO.getPartnerid() != users.getId()){
+                mav.setViewName("redirect:/" + Constants.Characters.BLANK);
+                return mav;
+            }
+            orderDAO.updatePurchaseOrder(true, id);
             mailTest.mailSend(orderDTO.getUserEmail(), MailContent.PURCHASED(orderDTO), "Thanh Toán Thành Công");
             String totalPriceAfter1 = orderDTO.getTotalPriceAfter();
             Integer totalPriceAfter = Integer.parseInt(totalPriceAfter1);
@@ -306,11 +310,17 @@ public class ManagementController {
     @GetMapping(Constants.Url.REFUND_PARTNER_ORDER)
     public ModelAndView refundOrder(HttpSession session, @PathVariable("id") Integer id) {
         ModelAndView mav = new ModelAndView();
-//        if (!CheckSession.partner(session)) {
-//            mav.setViewName("redirect:/" + Constants.Characters.BLANK);
-//            return mav;
-//        }
+        if (!CheckSession.partner(session)) {
+            mav.setViewName("redirect:/" + Constants.Characters.BLANK);
+            return mav;
+        }
         try {
+            OrderDTO orderDTO = orderDAO.getOrderDTO(id);
+            Users users = (Users) session.getAttribute(Constants.SessionKey.USER);
+            if(orderDTO.getPartnerid() != users.getId()){
+                mav.setViewName("redirect:/" + Constants.Characters.BLANK);
+                return mav;
+            }
             orderDAO.updatePurchaseOrder(false, id);
             mav.setViewName("redirect:" + Constants.Url.MANAGEMENT_URL + Constants.Url.MANAGEMENT_PARTNER_DELETE_ORDER_URL);
             return mav;
@@ -323,12 +333,17 @@ public class ManagementController {
     @GetMapping(Constants.Url.DELETE_ORDER_CUSTOMER)
     public ModelAndView deleteOrder(HttpSession session, @PathVariable("id") Integer id, @PathVariable("purchase") boolean purchase) {
         ModelAndView mav = new ModelAndView();
-        if (!CheckSession.partner(session)) {
+        if (!CheckSession.isLogin(session)) {
             mav.setViewName("redirect:/" + Constants.Characters.BLANK);
             return mav;
         }
         try {
-          OrderDTO orderDTO =  orderDAO.getOrderDTO(id);
+            OrderDTO orderDTO = orderDAO.getOrderDTO(id);
+            Users users = (Users) session.getAttribute(Constants.SessionKey.USER);
+            if(orderDTO.getPartnerid() != users.getId()){
+                mav.setViewName("redirect:/" + Constants.Characters.BLANK);
+                return mav;
+            }
           Date date = new Date();
             if(date.getTime() - orderDTO.getOrderDate().getTime() >  orderDTO.getPolicy()){
                 if (purchase == true) {
@@ -358,16 +373,21 @@ public class ManagementController {
     @PostMapping(Constants.Url.CHANGE_CUSTOMER_INFORMATION)
     public ModelAndView registerCustomer(HttpServletRequest request, @Validated @ModelAttribute(Constants.SessionKey.USER) Users users, BindingResult result, HttpSession session) {
         ModelAndView mav = new ModelAndView();
+        if(!CheckSession.isLogin(session)){
+            mav.setViewName("redirect:/" + Constants.Characters.BLANK);
+            return mav;
+        }
         if (result.hasErrors()) {
             mav.setViewName(INFOMATION_SCREEN);
             return mav;
         }
-        if(!CheckSession.admin(session)){
-            mav.setViewName("redirect:/" + Constants.Characters.BLANK);
-            return mav;
-        }
         try {
             Users us = user.getById(users.getId());
+            Users us1 = (Users) session.getAttribute(Constants.SessionKey.USER);
+            if(us.getId() != us1.getId()){
+                mav.setViewName("redirect:/" + Constants.Characters.BLANK);
+                return mav;
+            }
             CommonsMultipartFile[] files = users.getFileData();
             for (int i = 0; i < files.length; i++) {
                 CommonsMultipartFile file = files[i];
